# docs/swagger.yaml
definitions:
  model.Evaluation:
    properties:
      Id:
        type: integer
      Note:
        type: string
      Rating:
        type: integer
    type: object
  model.EvaluationAdd:
    properties:
      Note:
        type: string
      Rating:
        type: integer
    type: object
  model.EvaluationFullView:
    properties:
      Id:
        type: integer
      Note:
        type: string
      Rating:
        type: string
    type: object
  model.EvaluationPartialView:
    properties:
      Id:
        type: integer
      Rating:
        type: string
    type: object
  model.EvaluationUpd:
    properties:
      Note:
        type: string
      Rating:
        type: integer
    type: object
  model.Users:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: Autentica o usuário e gera o token para os próximos acessos
      parameters:
      - description: Do login
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/model.Users'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      summary: Realizar autenticação
    put:
      consumes:
      - application/json
      description: Atualiza o token de autenticação do usuário
      parameters:
      - description: Refresh token
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/model.Users'
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Users'
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      summary: Atualiza token de autenticação
  /evaluation:
    get:
      consumes:
      - application/json
      description: Exibe a lista, sem todos os campos, de todas as avaliações
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.EvaluationPartialView'
            type: array
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Recupera as avaliações
    post:
      consumes:
      - application/json
      description: Cria uma avaliação sobre a utilização da aplicação
      parameters:
      - description: Token
        in: header
        name: Bearer
        required: true
        type: string
      - description: Add evaluation
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/model.EvaluationAdd'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Evaluation'
        "400":
          description: Bad request
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Adicionar uma avaliação
  /evaluation/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui uma avaliação realizada
      operationId: get-string-by-int
      parameters:
      - description: Token
        in: header
        name: Bearer
        required: true
        type: string
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Evaluation'
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Exclui uma avaliação pelo ID
    get:
      consumes:
      - application/json
      description: Exibe os detalhes de uma avaliação pelo ID
      operationId: get-string-by-int
      parameters:
      - description: Token
        in: header
        name: Bearer
        required: true
        type: string
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EvaluationFullView'
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Recupera uma avaliação pelo id
    put:
      consumes:
      - application/json
      description: Atualiza uma avaliação sobre a utilização da aplicação
      parameters:
      - description: Token
        in: header
        name: Bearer
        required: true
        type: string
      - description: Udpdate evaluation
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/model.EvaluationUpd'
          type: object
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Evaluation'
        "400":
          description: Bad request
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Atualiza uma avaliação
swagger: "2.0"